Some compilers need a cast for LOG << size()
git-svn-id: svn://svn.code.sf.net/p/boomerang/code/trunk@2126 dc5e1367-85d4-4bd1-96b4-a20bfbfd92d5

+++ boomerang/db/proc.cpp
@@ -20,7 +20,7 @@
  *============================================================================*/
 
 /*
- * $Revision: 1.335 $	// 1.238.2.44
+ * $Revision: 1.336 $	// 1.238.2.44
  *
  * 14 Mar 02 - Mike: Fixed a problem caused with 16-bit pushes in richards2
  * 20 Apr 02 - Mike: Mods for boomerang
@@ -3548,8 +3548,8 @@ void UserProc::removeUnusedLocals() {
 		std::string& name = const_cast<std::string&>(it->first);
 		// LOG << "Considering local " << name << "\n";
 		if (VERBOSE && all && removes.size())
-			LOG << "WARNING: defineall seen in procedure " << name.c_str() << " so not removing " << removes.size() <<
-				" locals\n";
+			LOG << "WARNING: defineall seen in procedure " << name.c_str() << " so not removing " << (int)removes.size()
+				 << " locals\n";
 		if (usedLocals.find(name) == usedLocals.end() && !all) {
 			if (VERBOSE)
 				LOG << "removed unused local " << name.c_str() << "\n";
